Move searchResults query into custom hook.
