+++ packages/openneuro-app/src/scripts/refactor_2021/search/search-container.tsx
@@ -1,9 +1,10 @@
-import React, { FC, useContext } from 'react'
-import { SearchParamsCtx } from './search-params-ctx'
+import React, { FC } from 'react'
 import {
   SearchPage,
   sortBy,
   SearchSortContainerExample,
+  SearchResultsList,
+  Button,
 } from '@openneuro/components'
 import {
   KeywordInput,
@@ -21,10 +22,15 @@ import {
   StudyDomainSelect,
 } from './inputs'
 import FiltersBlockContainer from './filters-block-container'
-import SearchResultsContainer from './search-results-container'
+import { useCookies } from 'react-cookie'
+import { getUnexpiredProfile } from '../authentication/profile'
+import { useSearchResults } from './use-search-results'
 
 const SearchContainer: FC = () => {
-  const { searchParams, setSearchParams } = useContext(SearchParamsCtx)
+  const [cookies] = useCookies()
+  const profile = getUnexpiredProfile(cookies)
+  const { loading, data, fetchMore, refetch, variables, error } =
+    useSearchResults()
 
   return (
     <SearchPage
@@ -60,7 +66,23 @@ const SearchContainer: FC = () => {
           <StudyDomainSelect />
         </>
       )}
-      renderSearchResultsList={() => <SearchResultsContainer />}
+      renderSearchResultsList={() =>
+        loading ? (
+          <h1>Datasets loading placeholder</h1>
+        ) : (
+          <h1>
+            <SearchResultsList items={data?.datasets.edges} profile={profile} />
+            <div className="grid grid-nogutter" style={{ width: '100%' }}>
+              <div className="col col-12 results-count">
+                Showing <b>25</b> of <b>100</b> Datasets
+              </div>
+              <div className="col col-12 load-more ">
+                <Button label="Load More" />
+              </div>
+            </div>
+          </h1>
+        )
+      }
     />
   )
 }
