+++ packages/openneuro-app/src/scripts/refactor_2021/search/use-search-results.tsx
@@ -1,9 +1,7 @@
 import React, { FC, useContext, useEffect } from 'react'
 import { gql, useQuery } from '@apollo/client'
-import { useCookies } from 'react-cookie'
 import { SearchResultsList, Button } from '@openneuro/components'
 import { SearchParamsCtx } from './search-params-ctx'
-import { getUnexpiredProfile } from '../authentication/profile'
 
 const searchQuery = gql`
   query searchDatasets($q: String!, $cursor: String) {
@@ -84,41 +82,13 @@ const searchQuery = gql`
     }
   }
 `
-
-const SearchResultsContainer: FC = () => {
-  const [cookies] = useCookies()
-  const profile = getUnexpiredProfile(cookies)
+export const useSearchResults = () => {
   const { searchParams, setSearchParams } = useContext(SearchParamsCtx)
 
-  const {
-    loading,
-    data,
-    fetchMore,
-    refetch,
-    variables,
-    error = null,
-  } = useQuery(searchQuery, {
+  return useQuery(searchQuery, {
     variables: {
       q: 'FSL',
     },
     errorPolicy: 'ignore',
   })
-
-  return loading ? (
-    <h1>Datasets loading placeholder</h1>
-  ) : (
-    <h1>
-      <SearchResultsList items={data?.datasets.edges} profile={profile} />
-      <div className="grid grid-nogutter" style={{ width: '100%' }}>
-        <div className="col col-12 results-count">
-          Showing <b>25</b> of <b>100</b> Datasets
-        </div>
-        <div className="col col-12 load-more ">
-          <Button label="Load More" />
-        </div>
-      </div>
-    </h1>
-  )
 }
-
-export default SearchResultsContainer
